Add SetFitModel.to (#229)
* Add SetFitModel.to

* Add docstring for 'to' to clarify that we don't copy

Note that this does mean that we differ from e.g. torch, which creates a moved copy rather than moving the e.g. tensor itself to a different device.

* Add tests for 'to'.

If CUDA is available, then try to move a model from the CPU to the GPU and back to the CPU.

* Reformat test_modeling

Co-authored-by: Tom Aarsen <Cubiegamedev@gmail.com>